[BUILDFIX] support fmtlib 10
With fmtlib 10, fmt::format() does not handle WasmEdge::ErrCode as it is.
Pass this class to cast operator so that fmtlib 10 can handle this.

Closes #2646 .

+++ include/executor/engine/atomic.ipp
@@ -42,7 +42,7 @@ TypeT<T> Executor::runAtomicWaitOp(Runtime::StackManager &StackMgr,
 
   if (auto Res = atomicWait<T>(MemInst, Address, RawValue.get<T>(), Timeout);
       unlikely(!Res)) {
-    spdlog::error(Res.error());
+    spdlog::error(static_cast<uint32_t>(Res.error()));
     spdlog::error(
         ErrInfo::InfoInstruction(Instr.getOpCode(), Instr.getOffset()));
     return Unexpect(Res);
